Removed method to load into DSS using in-place copy
Removed method upload_cloud_file from base loader, which performed
a load by copy into the DSS and bypassed the typical copy to staging
by tagging the original file in place.

All loading of cloud data files into the DSS should be done by reference.
In part, this is due to the potential/pending of removal of support for
`directurls` from the HCA DSS.

The use case for which this method was originally added is no longer
applicable/important.