Added `message` function to use in tests
Instead of simply placing comments in unit tests, pass that same text
to the `message` function to provide timestamped messages as the
test is executing. This provides timely output about what the test is
doing and the timestamp can be helpful when debugging to correlate
test activity with timestamped log messages from a server/service
with which the test is interacting.