Update loader to work with whitelisted dss (resolves STOR-2) (#37)
This should have just worked out of the box without this PR but there were some bugs to resolve.

bug 1.
The integration tests used the requests library to get some stuff. Once things become whitelisted and authenticated this becomes a problem. The fix was easy, just use the dss stuff so it all happens correctly behind the scene.

bug 2.
The way the dss config was initialized caused some problems. The config remembers your setup from before and will automatically pull configurations depending on the name of the config. HCAConfig forced the name to always be "hca" which meant that if ever you had run the command before (no matter when you called HCAConfig() things would already be initialized in an "hca"-y way and it would be unable to point to any other dss instance.

The easy solution---you say---just initialize with name not as "hca", like this. It's a great idea, but the HCAConfig code is written in such a way that doing this causes an error. the culpret is line 9 where name is passed to super()__init__ preventing us from overwriting it later down the inheritance tree.

To work around this I did a very hacky monkey patch to prevent setting this default name="hca". This works great for our purposes; any previous "hca" configs are left untouched if the user had made some before, and we make fresh ones for ourselves.

Fixes #29, #35